% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_alignment_similarity.R
\name{aggregate_alignment_similarity}
\alias{aggregate_alignment_similarity}
\title{Summarize Similarities After Factor Alignment}
\usage{
aggregate_alignment_similarity(S, alignment)
}
\arguments{
\item{S}{A square similarity matrix used to obtain the factor
alignment.}

\item{alignment}{An alignment object returned by
\code{factor_alignment_hungarian()}.}
}
\value{
A list containing:
\describe{
\item{\code{pair_similarities}}{
Similarities for each matched pair of latent factors.
}
\item{\code{mean_similarity}}{
Mean similarity across the matched factor pairs.
}
\item{\code{sum_similarity}}{
Sum of the similarities across the matched factor pairs.
}
}
}
\description{
Extracts the similarities associated with an optimal factor
alignment and calculates their mean and total similarity.
}
\seealso{
\code{\link{factor_alignment_hungarian}}
}
