% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_H_samples_bbmf.R
\name{align_H_samples_bbmf}
\alias{align_H_samples_bbmf}
\title{Align Posterior Samples of H to a Reference Matrix}
\usage{
align_H_samples_bbmf(H_ref, H_samples)
}
\arguments{
\item{H_ref}{A binary reference matrix of dimension
\eqn{R \times G}.}

\item{H_samples}{A three-dimensional binary array of dimension
\eqn{R \times G \times T}.}
}
\value{
A three-dimensional array with the same dimensions as
\code{H_samples}, with factors aligned to \code{H_ref}.
}
\description{
Aligns the rows of each posterior sample of the BBMF factor-feature
matrix H to a reference factor matrix using Jaccard similarity and
the Hungarian algorithm.
}
\details{
For each posterior sample, pairwise Jaccard similarities are computed
between the rows of \code{H_ref} and the sampled H matrix. The
Hungarian algorithm is then used to determine the optimal factor
permutation.
}
\seealso{
\code{\link{jaccard_rows}},
\code{\link{factor_alignment_hungarian}}
}
