% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_cols.R
\name{jaccard_cols}
\alias{jaccard_cols}
\title{Compute Jaccard Similarities Between Matrix Columns}
\usage{
jaccard_cols(X1, X2)
}
\arguments{
\item{X1}{A binary matrix with samples represented by rows.}

\item{X2}{A binary matrix with the same number of rows as
\code{X1}.}
}
\value{
A numeric matrix of dimension
\code{ncol(X1)} by \code{ncol(X2)}. Entry \eqn{(i,j)} contains
the Jaccard similarity between column \eqn{i} of \code{X1} and
column \eqn{j} of \code{X2}.
}
\description{
Computes all pairwise Jaccard similarities between the columns of
two binary matrices having the same number of rows.
}
\details{
For binary vectors \eqn{x} and \eqn{y}, the Jaccard similarity is

\deqn{
J(x,y) =
\frac{|x \cap y|}{|x \cup y|}.
}

If both vectors contain only zeros, their union is empty and the
similarity is defined here to be zero.

This function is particularly useful for comparing columns of two
sample-factor matrices \code{W}.
}
