% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_rows.R
\name{jaccard_rows}
\alias{jaccard_rows}
\title{Compute Jaccard Similarities Between Matrix Rows}
\usage{
jaccard_rows(X1, X2)
}
\arguments{
\item{X1}{A binary matrix with features represented by columns.}

\item{X2}{A binary matrix with the same number of columns as
\code{X1}.}
}
\value{
A numeric matrix of dimension
\code{nrow(X1)} by \code{nrow(X2)}. Entry \eqn{(i,j)} contains
the Jaccard similarity between row \eqn{i} of \code{X1} and row
\eqn{j} of \code{X2}.
}
\description{
Computes all pairwise Jaccard similarities between the rows of two
binary matrices having the same number of columns.
}
\details{
For binary vectors \eqn{x} and \eqn{y}, the Jaccard similarity is

\deqn{
J(x,y) =
\frac{|x \cap y|}{|x \cup y|}.
}

If both vectors contain only zeros, their union is empty and the
similarity is defined here to be zero.

This function is particularly useful for comparing rows of two
factor-feature matrices \code{H}.
}
