% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_X_one_chain_bbmf.R
\name{reconstruct_X_one_chain_bbmf}
\alias{reconstruct_X_one_chain_bbmf}
\title{Reconstruct Binary Matrices Across One BBMF Chain}
\usage{
reconstruct_X_one_chain_bbmf(W_samples, H_samples)
}
\arguments{
\item{W_samples}{A three-dimensional binary array of dimension
\eqn{K \times R \times T}, containing posterior samples of
the sample-factor matrix \eqn{W}.}

\item{H_samples}{A three-dimensional binary array of dimension
\eqn{R \times G \times T}, containing posterior samples of
the factor-feature matrix \eqn{H}.}
}
\value{
A three-dimensional binary array of dimension
\eqn{K \times G \times T}. Slice \code{t} contains the Boolean
reconstruction obtained from the corresponding posterior samples
of \eqn{W} and \eqn{H}.
}
\description{
Reconstructs the binary matrix \eqn{X} for every posterior sample
of \eqn{W} and \eqn{H} from a single BBMF MCMC chain.
}
\details{
For each posterior iteration \eqn{t}, the reconstruction is

\deqn{
\widetilde{X}^{(t)}
=
W^{(t)} \circ H^{(t)},
}

where \eqn{\circ} denotes the Boolean matrix product implemented
by \code{bool_prod()}.
}
\examples{
\dontrun{
X_samples <- reconstruct_X_one_chain_bbmf(
  W_samples = Ws[[1]],
  H_samples = Hs[[1]]
)
}

}
\seealso{
\code{\link{bool_prod}}
}
