% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_H_posterior_bbmf.R
\name{summarize_H_posterior_bbmf}
\alias{summarize_H_posterior_bbmf}
\title{Summarize Posterior BBMF Factor Loadings}
\usage{
summarize_H_posterior_bbmf(H_chains, H_ref)
}
\arguments{
\item{H_chains}{A list of aligned three-dimensional binary arrays.
Each array must have dimension \eqn{R \times G \times T}.}

\item{H_ref}{A reference binary matrix of dimension
\eqn{R \times G}, used to provide factor and feature names.}
}
\value{
A list containing:
\describe{
\item{\code{H_all}}{
Posterior samples of H combined across chains.
}
\item{\code{probability}}{
Posterior inclusion probability for each factor-feature entry.
}
\item{\code{uncertainty}}{
Scaled posterior uncertainty for each factor-feature entry.
}
}
}
\description{
Combines aligned posterior samples of the BBMF factor-feature matrix
H across MCMC chains and computes posterior inclusion probabilities
and uncertainty scores.
}
\details{
For factor \eqn{r} and feature \eqn{g}, the posterior inclusion
probability is

\deqn{
\widehat{P}_{rg}
=
\frac{1}{M}
\sum_{m=1}^{M}
H_{rg}^{(m)}.
}

Posterior uncertainty is calculated as

\deqn{
U_{rg}
=
4\widehat{P}_{rg}
(1-\widehat{P}_{rg}).
}
}
